making new class
