+++ strawberryfields/apps/data/SampleData/MUTAG_0.npz
