+++ strawberryfields/apps/data/SampleData/MUTAG_0_A.npz
