+++ strawberryfields/apps/data/SampleData/MUTAG_1.npz
