+++ strawberryfields/apps/data/SampleData/MUTAG_1_A.npz
