+++ strawberryfields/apps/data/SampleData/MUTAG_2.npz
