+++ strawberryfields/apps/data/SampleData/MUTAG_2_A.npz
