+++ strawberryfields/apps/data/SampleData/MUTAG_3.npz
