+++ strawberryfields/apps/data/SampleData/MUTAG_3_A.npz
