+++ strawberryfields/apps/data/SampleData/formic.npz
