+++ strawberryfields/apps/data/SampleData/formic_Ud.npz
