+++ strawberryfields/apps/data/SampleData/formic_delta.npz
