+++ strawberryfields/apps/data/SampleData/formic_w.npz
