+++ strawberryfields/apps/data/SampleData/formic_wp.npz
