+++ strawberryfields/apps/data/SampleData/p_hat300-1.npz
