+++ strawberryfields/apps/data/SampleData/p_hat300-1_A.npz
