+++ strawberryfields/apps/data/SampleData/planted.npz
