+++ strawberryfields/apps/data/SampleData/planted_A.npz
